requirements.txt
google-api-python-client==2.131.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
anthropic>=0.40.0
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
tqdm==4.66.4
tabulate==0.9.0
